requirements/staging.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.0.2
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.3
django-cors-headers==2.1.0
django-extensions==1.8.1
django-model-utils==3.0.0
djangorestframework==3.6.3
geopy==1.11.0
idna==2.5
ipdb==0.10.2
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
kombu==4.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.1
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.0
git+https://github.com/davidchua/pymessenger.git
PySocks==1.6.7
pytz==2017.2
requests==2.18.1
requests-toolbelt==0.8.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
typing==3.6.1
urllib3==1.21.1
vine==1.1.4
wcwidth==0.1.7
boto==2.48.0
django-ses==0.8.2
channels==1.1.8
wit==4.3.0
asgi-redis==1.4.3
service-identity==17.0.0
pandas==0.21.0
openpyxl==2.4.9
pyenchant==1.6.11
textblob==0.13.0
wikipedia==1.4.0
